+++ webapp/src/Hooks.cpp
@@ -247,7 +247,7 @@ void DestroyDatabase(Webapp* app, Database* db) {
 	return app->DestroyDatabase(db);
 }
 
-Database* GetDatabase(Webapp* app, uint64_t index) {
+Database* GetDatabase(Webapp* app, size_t index) {
 	if(app == NULL) return NULL;
 	return app->GetDatabase(index);
 }
